[Llava] Add csv image loading in C++ runner
Rationale - We don't want to depend on Torch when building for Android.

This add two things,
(1) for AoT, python image_util optionally generates a .csv from a jpg. In
addition to .pt.

(2) add a runtime runner flag which hints at the provided image is a
csv. And if so, the runner parses the csv and feeds it to the model.

This is very naive and obviously fragile. Added some checks in python.

Tested few ways,
- On M1, with torch, loaded both .pt and .csv generated from the same
  jpg. And the LLM produces same text.
- On Android, without torch, loaded .csv and it also produces similar
  text.:wq