[ET-VK] Introduce `virtual_transpose()` to `vTensor` for no copy transposition
## Context

With `axis_map` integrated into matrix multiplication, we can now test no-copy transposes for texture backed tensors.

Transposing a tensor can be done without modifying the storage by swapping elements in the tensor's `axis_map`, and also updating the layout of the tensor if the packed dimension was one of the dims that were transposed.

Differential Revision: [D62652009](https://our.internmc.facebook.com/intern/diff/D62652009/)

ghstack-source-id: 242452080
Pull Request resolved: https://github.com/pytorch/executorch/pull/5353